% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compound_interest.R
\name{compounding_interest}
\alias{compounding_interest}
\title{Title Compound Interest}
\usage{
compounding_interest(p, r, t)
}
\arguments{
\item{p}{numeric}

\item{r}{numeric}

\item{t}{numeric}
}
\value{
numeric
}
\description{
Title Compound Interest
}
\examples{
compounding_interest(6, 0.0425, 6) # 1283.68
}
