% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_value.R
\name{future_value}
\alias{future_value}
\title{Title Future Value}
\usage{
future_value(p, r, t)
}
\arguments{
\item{p}{numeric}

\item{r}{numeric}

\item{t}{numeric}
}
\value{
numeric
}
\description{
Title Future Value
}
\examples{
future_value(100,5,1)
}
